Refactor client API
The RPC type is replaced with a more mockable API type.
It's shared between components using a provider/consumer
interface similar to the upcoming React context API. The
config data has been split from the API as a first step
towards a state-based interface.

Another intent here is that the API module can be used
on the server for all data loading, eventually. This will
make it easier to handle SSR.